ACTIVITY-AGENT-TRACKER-GROUP-6/requirements.txt
Flask==2.1.2
boto3==1.20.0
Pillow==9.1.0
pyautogui==0.9.53
pynput==1.7.4
psutil==5.9.0
